ShutNet v3.2

+++ ShutNet.ps1
@@ -14,11 +14,15 @@ Param (
 )
 
 foreach ($ip in $IPAddress) {
-    $computerName = (Resolve-DnsName -Name $ip -ErrorAction SilentlyContinue).NameHost
+    if (Test-Connection -ComputerName $ip -Count 1 -ErrorAction SilentlyContinue) {
+        $computerName = (Resolve-DnsName -Name $ip -ErrorAction SilentlyContinue).NameHost
 
-    if ($computerName) {
-        Stop-Computer -ComputerName $computerName -Force -Confirm:$false
+        if ($computerName) {
+            Stop-Computer -ComputerName $computerName -Force -Confirm:$false
+        } else {
+            Write-Warning "Could not resolve computer name for IP address $ip"
+        }
     } else {
-        Write-Warning "Could not resolve computer name for IP address $ip"
+        Write-Warning "Could not connect to IP address $ip"
     }
 }
